move fullscreen and noSleep to service
